backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
google-cloud-datastore==2.1.0
gunicorn==20.1.0